Добавлена цепочка селекторов и поддержка JSON
